Improvement/conversion flag (#179)
* Added -Wconversion flag to Makefile and cleaned up conversions in src/main/seal.c

* Cleaned up type conversions in src/cipher/aes_gcm.c

* Cleaned up type conversions in src/cipher/cipher.c

* Cleaned up type conversions in src/cipher/aes_keywrap_5649pad.c

* Cleaned up type conversions in src/cipher/aes_keywrap_3394nopad.c

* Cleaned up type conversions in src/network/tls_util.c

* Cleaned up type conversions in src/protocol/kmip_util.c and src/protocol/nsl_util.c

* Cleaned up type conversions in src/tpm/object_tools.c

* Cleaned up type conversions in src/tpm/pcrs.c

* Cleaned up type conversions in src/tpm/kmyth_seal_unseal_impl.c.

* Cleaned up type conversions in src/tpm/marshalling_tools.c

* Cleaned up type conversions in src/tpm/storage_key_tools.c

* Cleaned up type conversions in src/tpm/tpm2_interface.c and documented a constraint on KMYTH_DIGEST_SIZE in include/defines.h

* Cleaned up type conversions in utils/src/memory_util.c

* Cleaned up type conversions in utils/src/file_io.c

* Cleaned up type conversions in utils/src/formatting_tools.c

* Cleaned up type conversions in logger/src/kmyth_log.c

* Cleaned up type conversions in src/main/reseal.c

* Cleaned up type conversions in src/main/nsl_client.c

* Cleaned up type conversions in src/main/nsl_server.c

* Cleaned up type conversions in test/src/cipher/aes_keywrap_test.c

* Cleaned up type conversions in test/src/cipher/aes_gcm_test.c

* Cleaned type conversions in test/src/cipher/cipher_test.c

* Cleaned up type conversions in sgx/common/src/retrieve_key_protocol.c

* Fixed a bunch of interlinked type conversion issues.

* Fixed inconsistency between declaration and definition of setup_socket_ocall.

* Cleaned up type conversions in sgx/untrusted/src/util/msg_util.c

* Cleaned up type conversions in sgx/common/src/ecdh_util.c

* Cleaned up type conversions in test/src/cpm/formatting_tools_test.c

+++ logger/src/kmyth_log.c
@@ -298,7 +298,7 @@ void log_event(const char *src_file,
   va_list args;
 
   va_start(args, message);
-  vsnprintf(out, log_settings.applog_max_msg_len + 1, message, args);
+  vsnprintf(out, (size_t)log_settings.applog_max_msg_len + 1, message, args);
   va_end(args);
 
   // force severity to a valid value by masking (only use three lowest bits)
